Review counting time elapsed, scaffolding db operations

+++ Flashcards.Core/Models/DailyActivity.cs
@@ -24,7 +24,7 @@ namespace Flashcards.Core.Models
         public int ReviewedFlashcardsCount { get; set; }
 
         [Required]
-        public int MinutesSpentLearning { get; set; }
+        public double MinutesSpentLearning { get; set; }
 
         [Required]
         [ForeignKey("User")]
