+++ Flashcards.Core/Models/DeckActivity.cs
@@ -21,7 +21,7 @@ namespace Flashcards.Core.Models
         public int ReviewedFlashcardsCount { get; set; }
 
         [Required]
-        public int MinutesSpentLearning { get; set; }
+        public double MinutesSpentLearning { get; set; }
 
         [Required]
         [ForeignKey("Deck")]
